{{cookiecutter.project_slug}}/backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.108.0,<0.109.0
uvicorn[standard]>=0.25.0,<0.26.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database
sqlalchemy>=2.0.25,<3.0.0
asyncpg>=0.29.0,<0.30.0
alembic>=1.13.0,<2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-multipart>=0.0.6,<0.1.0

# Caching & Session Management
redis>=5.0.1,<6.0.0

# CrewAI & AI
crewai>=0.14.0,<0.15.0
openai>=1.6.0,<2.0.0
anthropic>=0.8.0,<0.9.0

# HTTP Client
httpx>=0.26.0,<0.27.0
aiohttp>=3.9.0,<4.0.0

# Background Tasks
celery>=5.3.0,<6.0.0

# Monitoring & Logging
python-json-logger>=2.0.0,<3.0.0
structlog>=23.2.0,<24.0.0

# System Monitoring
psutil>=5.9.0,<6.0.0

# Web Scraping and Parsing
beautifulsoup4>=4.12.0,<5.0.0
lxml>=4.9.0,<5.0.0

# Email
aiosmtplib>=3.0.0,<4.0.0
jinja2>=3.1.0,<4.0.0

# Validation & Serialization
email-validator>=2.1.0,<3.0.0

# Development & Testing
pytest>=7.4.4,<8.0.0
pytest-asyncio>=0.23.2,<0.24.0
pytest-cov>=4.1.0,<5.0.0
pytest-mock>=3.12.0,<4.0.0
httpx>=0.26.0,<0.27.0

# Code Quality
black>=23.12.1,<24.0.0
flake8>=7.0.0,<8.0.0
isort>=5.13.2,<6.0.0
mypy>=1.8.0,<2.0.0

# Security
bandit>=1.7.5,<2.0.0
safety>=2.3.5,<3.0.0

# Pre-commit hooks
pre-commit>=3.6.0,<4.0.0